+++ modules/sd_offload.py
@@ -14,7 +14,7 @@ from modules.timer import process as process_timer
 debug = os.environ.get('SD_MOVE_DEBUG', None) is not None
 verbose = os.environ.get('SD_MOVE_VERBOSE', None) is not None
 debug_move = log.trace if debug else lambda *args, **kwargs: None
-offload_warn = ['sc', 'sd3', 'f1', 'f2', 'h1', 'hunyuandit', 'auraflow', 'omnigen', 'omnigen2', 'cogview4', 'cosmos', 'chroma', 'x-omni', 'hunyuanimage', 'hunyuanimage3', 'longcat']
+offload_allow_none = ['sd', 'sdxl']
 offload_post = ['h1']
 offload_hook_instance = None
 balanced_offload_exclude = ['CogView4Pipeline', 'MeissonicPipeline']
@@ -132,7 +132,7 @@ def apply_sequential_offload(sd_model, op:str='model', quiet:bool=False):
 
 
 def apply_none_offload(sd_model, op:str='model', quiet:bool=False):
-    if shared.sd_model_type in offload_warn or 'video' in shared.sd_model_type:
+    if shared.sd_model_type not in offload_allow_none:
         shared.log.warning(f'Setting {op}: offload={shared.opts.diffusers_offload_mode} type={shared.sd_model.__class__.__name__} large model')
     else:
         shared.log.quiet(quiet, f'Setting {op}: offload={shared.opts.diffusers_offload_mode} limit={shared.opts.cuda_mem_fraction}')